程序测试/requirements.txt
pyserial 
tkinter 
opencv-python 
pyzbar